- Use Express for routing - Move resolvers to src/graphql/schema.js - Removing nodemon due to broken dependencies - install apollo-server-express and express - Remove empty /index.js file
